Move common types from bifrost to restate-types
Needed for future PRs. Mostly mechanical.

Test Plan:
Tests...
